requirements.txt
pygame
kivy
twisted
pytest
pytest-xdist
lxml
coverage
coveralls
pillow